+++ lib/about.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class AboutPage extends StatefulWidget {
+  const AboutPage({super.key});
+
+  @override
+  State<AboutPage> createState() => _AboutPageState();
+}
+
+class _AboutPageState extends State<AboutPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: Text('About Page'),
+      ),
+    );
+  }
+}
+
