+++ lib/products.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class ProductsPage extends StatefulWidget {
+  const ProductsPage({super.key});
+
+  @override
+  State<ProductsPage> createState() => _ProductsPageState();
+}
+
+class _ProductsPageState extends State<ProductsPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: Text('Products Page'),
+      ),
+    );
+  }
+}
+
