+++ lib/projects.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class ProjectsPage extends StatefulWidget {
+  const ProjectsPage({super.key});
+
+  @override
+  State<ProjectsPage> createState() => _ProjectsPageState();
+}
+
+class _ProjectsPageState extends State<ProjectsPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: Center(
+        child: Text('Projects Page'),
+      ),
+    );
+  }
+}
+
